requirements.txt
setuptools
numpy
pandas
scikit-learn
seaborn
python-dotenv
mysql-connector-python
pymysql
# -e .